requirements.txt
paho-mqtt
certifi
mysql-connector-python
sqlalchemy
